+++ Source/Chronometer/IMemoryOptimizer.cs
@@ -1,0 +1,13 @@
+﻿namespace Narkhedegs.PerformanceMeasurement
+{
+    /// <summary>
+    /// Provides methods to optimize memory of the current process.
+    /// </summary>
+    internal interface IMemoryOptimizer
+    {
+        /// <summary>
+        /// Optimizes the memory of the current process.
+        /// </summary>
+        void Optimize();
+    }
+}