+++ ReleaseNotes.md
@@ -1,3 +1,9 @@
+### New in 0.0.3 (Released 2015/08/27)
+* Re-factor Chronometer class to move memory optimization functionality into its own class
+* Re-factor Chronometer class to move performance optimization functionality into its own class
+* Add functionality to detect debug mode and prevent measurement under debug mode by default
+* Add AllowMeasurementsUnderDebugMode option to ChronometerOptions to override the default behaviour in debug mode
+
 ### New in 0.0.2 (Released 2015/08/19)
 * Change NuGet package id to Narkhedegs.Chronometer